Only apply to v2
